Refactor additional component UI

+++ src/main/java/games/rednblack/editor/view/ui/properties/UIRemovableProperties.java
@@ -4,9 +4,12 @@ import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.kotcrab.vis.ui.VisUI;
+import com.kotcrab.vis.ui.util.dialog.Dialogs;
 import com.kotcrab.vis.ui.widget.VisImageButton;
 import com.kotcrab.vis.ui.widget.VisTable;
 import games.rednblack.editor.utils.StandardWidgetsFactory;
+import games.rednblack.editor.view.stage.Sandbox;
+import games.rednblack.h2d.common.MsgAPI;
 
 /**
  * Created by CyberJoe on 7/2/2015.
@@ -21,11 +24,11 @@ public abstract class UIRemovableProperties extends UIItemCollapsibleProperties 
     public Table crateHeaderTable() {
         VisTable header = new VisTable();
         header.setBackground(VisUI.getSkin().getDrawable("expandable-properties-active-bg"));
-        VisImageButton collapseButton = new VisImageButton("expandable-properties-button");
-        VisImageButton closeButton = new VisImageButton("close-panel");
+        VisImageButton collapseButton = StandardWidgetsFactory.createImageButton("expandable-properties-button");
+        VisImageButton closeButton = StandardWidgetsFactory.createImageButton("close-properties");
         header.add(closeButton).left().padLeft(2);
         header.add(StandardWidgetsFactory.createLabel(title)).left().expandX().padLeft(6);
-        header.add(collapseButton).right().padRight(3);
+        header.add(collapseButton).right().padRight(8);
         header.addListener(new ClickListener() {
             @Override
             public void clicked(InputEvent event, float x, float y) {
@@ -38,8 +41,14 @@ public abstract class UIRemovableProperties extends UIItemCollapsibleProperties 
             @Override
             public void clicked(InputEvent event, float x, float y) {
                 super.clicked(event, x, y);
-                onRemove();
-                remove();
+                Dialogs.showConfirmDialog(Sandbox.getInstance().getUIStage(),
+                        "Delete Component", "Do you want to delete this component?",
+                        new String[]{"Cancel", "Delete"}, new Integer[]{0, 1}, r -> {
+                            if (r == 1) {
+                                onRemove();
+                                remove();
+                            }
+                        }).padBottom(20).pack();
             }
         });
         return header;
